+++ Adjust/adjust/src/main/java/com/adjust/sdk/ISdkClickHandler.java
@@ -4,7 +4,7 @@ package com.adjust.sdk;
  * Created by pfms on 31/03/16.
  */
 public interface ISdkClickHandler {
-    void init(boolean startsSending);
+    void init(IActivityHandler activityHandler, boolean startsSending);
     void pauseSending();
     void resumeSending();
     void sendSdkClick(ActivityPackage sdkClick);